+++ lib/AddListings/add_listing.dart
@@ -2,7 +2,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import 'color_scheme.dart';
+import '../Utils/color_scheme.dart';
 
 class AddListing extends StatefulWidget {
   const AddListing({Key? key}) : super(key: key);
